build: Added react-router-dom and create paths to some components
