+++ src/components/OnePage.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+import Banner from './OnePage/Banner';
+import AboutUs from './OnePage/AboutUs';
+import ActualEvents from './OnePage/ActualEvents';
+import WhyWorth from './OnePage/WhyWorth';
+import Contact from './OnePage/Contact';
+import Footer from './OnePage/Footer';
+
+const OnePage = () => {
+  return (
+    <>
+      <Banner />
+      <AboutUs />
+      <ActualEvents />
+      <WhyWorth />
+      <Contact />
+      <Footer />
+    </>
+  );
+};
+
+export default OnePage;
