+++ src/components/Registeration/Operation.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Operation = () => {
+  return <div>Rejestracja krok 1</div>;
+};
+
+export default Operation;
